# Down the rabbit hole with auto-serialization for Enumerations