...if you go chasing enumerations, and you know you're going to fall...

(for posterity)

@htmldoug