@vito-c 2020-04-09
~don't go chasing enumerations stick to the case classes and objects that you're used to~