@plokhotnyuk 2020-04-09
@russellremple `withName` is not the best option to resolve enum values by name. It uses iterator over an immutable bit set to scan through the set of enums and do a lot of boxing of `Int` values.

Bellow is a reverse flame graph that was built for results of the benchmark that use [the same method](https://github.com/plokhotnyuk/jsoniter-scala/blob/280cd170b23a1d0ccdac7fd481d16a2e21b8b3ef/jsoniter-scala-benchmark/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/benchmark/WeePickleFromTos.scala#L45) to resolve enums.

![image](https://user-images.githubusercontent.com/890289/78860285-1b1c1480-7a32-11ea-8a55-16ebaa313e73.png)

Moreover, an overhead of boxing will increase if the size of the enum set will grow.