@plokhotnyuk 2020-04-10
It happened that using `get` and `put` instead of `computeIfAbsent` shows better performance.
Please see difference in results for [this](https://github.com/plokhotnyuk/jsoniter-scala/commit/c7b48d5ddab3c0973d24ea87806eb0901770c5c6) commit.

##Before
```
[info] Benchmark                             (size)   Mode  Cnt       Score      Error  Units
[info] ArrayOfEnumsReading.avSystemGenCodec     128  thrpt    5  125417.028 ± 7978.193  ops/s
[info] ArrayOfEnumsReading.borerJson            128  thrpt    5   92371.028 ± 3239.216  ops/s
[info] ArrayOfEnumsReading.jacksonScala         128  thrpt    5  137093.637 ± 4876.654  ops/s
[info] ArrayOfEnumsReading.sprayJson            128  thrpt    5   61673.755 ± 1210.588  ops/s
[info] ArrayOfEnumsReading.uPickle              128  thrpt    5  108992.872 ±  559.195  ops/s
[info] ArrayOfEnumsReading.weePickle            128  thrpt    5  119829.647 ± 1483.497  ops/s
```

##After
```
[info] Benchmark                             (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfEnumsReading.avSystemGenCodec     128  thrpt    5  147601.012 ± 11644.643  ops/s
[info] ArrayOfEnumsReading.borerJson            128  thrpt    5  155107.924 ±  8589.481  ops/s
[info] ArrayOfEnumsReading.jacksonScala         128  thrpt    5  164926.548 ±   928.293  ops/s
[info] ArrayOfEnumsReading.sprayJson            128  thrpt    5   76496.355 ±  1291.351  ops/s
[info] ArrayOfEnumsReading.uPickle              128  thrpt    5  144977.867 ±  1580.067  ops/s
[info] ArrayOfEnumsReading.weePickle            128  thrpt    5  143804.008 ±   917.731  ops/s
```