@htmldoug 2020-04-10
Questions:
1. Will this compose, e.g. with `Option`, or would we need a separate solution for that?
2. If I have a `FromTo[Suit.Value]` in scope, would that take precedence over the macro implementation?