@russellremple 2020-04-10
As you expected, the answers are "no" and "no". That is the nice thing about implicits -- they compose. This is just for the common case where a data element is an enumeration (and not even an Enumeratum enumeration). So pretty narrow. And there is no precedence -- if it is an Enumeration, it gets this treatment, otherwise we look for an implicit. I wonder if the behavior could be defined _as_ an implicit rather than trying to go around implicit resolution, as long as it could be defined with a lower priority, then the answers would be "yes" and "yes". Thoughts?