@htmldoug 2020-04-10
>  I wonder if the behavior could be defined as an implicit rather than trying to go around implicit resolution

That would fit in better with the library. The expectation is that when you define a custom type, you toss a `FromTo[T]` in the companion object. `Enumeration.Value` is a custom type and I'd expect to create a custom pickler. Haoyi also suggested something to that effect in https://github.com/lihaoyi/upickle/issues/147#issuecomment-213767487.